+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [unreleased]
+
+- Added support for custom markup delimiters.
+
 ## [0.3.0] - 25-05-29
 
 - Fixed static analysis of lambda expressions (arrow functions). Previously we were not including lambda parameters in the scope of the expression. See [#12](https://github.com/jg-rp/ruby-liquid2/issues/12).